subjects: renamed bodc to nvs
