+++ invenio_vocabularies/contrib/subjects/config.py
@@ -36,8 +36,8 @@ euroscivoc_file_url = LocalProxy(
     lambda: current_app.config["VOCABULARIES_SUBJECTS_EUROSCIVOC_FILE_URL"]
 )
 
-bodc_puv_file_url = LocalProxy(
-    lambda: current_app.config["VOCABULARIES_SUBJECTS_BODC_PUV_FILE_URL"]
+nvs_file_url = LocalProxy(
+    lambda: current_app.config["VOCABULARIES_SUBJECTS_NVS_FILE_URL"]
 )
