+++ invenio_vocabularies/contrib/subjects/nvs/__init__.py
@@ -6,4 +6,4 @@
 # modify it under the terms of the MIT License; see LICENSE file for more
 # details.
 
-"""BODC Subjects module."""
+"""NVS Subjects module."""
